+++ desktop/electron-builder.yaml
@@ -1,0 +1,47 @@
+appId: app.dpkit
+copyright: Copyright © 2025 ${author}
+artifactName: dpkit-desktop-${os}-${version}.${ext}
+
+# General
+
+# publish:
+#   provider: github
+#   owner: datisthq
+#   releaseType: draft
+files:
+  - "build/client/**/*"
+  - "build/desktop/**/*"
+extraResources:
+  - from: "build/server"
+    to: "server"
+
+# Linux
+
+linux:
+  maintainer: datist.io
+  category: Utility
+  icon: icon.icns
+  target:
+    - AppImage
+    - deb
+
+# Macos
+# Some resources to understand values and github secrets
+# https://kilianvalkhof.com/2019/electron/notarizing-your-electron-application/
+# https://www.electron.build/code-signing
+# https://www.electron.build/app-builder-lib.interface.macconfiguration#notarize
+# https://github.com/electron/notarize
+mac:
+  hardenedRuntime: true
+  entitlements: "entitlements.mac.plist"
+  entitlementsInherit: "entitlements.mac.plist"
+
+# Windows
+# Variable ${name} doesn't work on windows
+
+win:
+  executableName: dpkit
+nsis:
+  shortcutName: ${productName}
+  uninstallDisplayName: ${productName}
+  createDesktopShortcut: always
