+++ desktop/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@dpkit/desktop",
+  "type": "module",
+  "version": "0.0.0-dev",
+  "private": true,
+  "scripts": {
+    "build": "tsc",
+    "start": "electron-vite dev"
+  },
+  "devDependencies": {
+    "@dpkit/browser": "workspace:*",
+    "@electron-toolkit/utils": "4.0.0",
+    "@loglayer/transport-electron-log": "2.1.10",
+    "electron-log": "5.4.3",
+    "loglayer": "6.9.1",
+    "electron-vite": "4.0.1",
+    "electron": "39.1.2",
+    "electron-builder": "26.0.12",
+    "electron-updater": "6.6.2"
+  }
+}
