+++ desktop/services/logger.ts
