+++ pnpm-workspace.yaml
@@ -4,6 +4,7 @@ packages:
   - browser
   - database
   - dataset
+  - desktop
   - document
   - dpkit
   - image
